refactor : Activities 엔드포인트 관련 쿼리 훅 useActivities로 통합 -> 메인페이지, 상세페이지
