+++ src/api/activities.ts
@@ -4,7 +4,6 @@ import { GetActivitiesParams } from "@/types/getActivitiesParams";
 
 const getActivities = async (params: GetActivitiesParams) => {
   const response = await client.get(API_ENDPOINTS.ACTIVITIES.LIST, { params });
-
   return response.data;
 };
